pyproject.toml
[project]
name = "agentic-ai-bootcamp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "dataclasses>=0.8",
    "dotenv>=0.9.9",
    "email-validator>=2.2.0",
    "gradio>=5.29.1",
    "ipykernel>=6.29.5",
    "ipython>=9.2.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "langsmith>=0.3.45",
    "lxml>=6.0.1",
    "nest-asyncio>=1.6.0",
    "openai>=1.78.1",
    "openai-agents>=0.0.15",
    "playwright>=1.55.0",
    "polygon-api-client>=1.16.1",
    "pydantic>=2.11.4",
    "pypdf>=5.5.0",
    "requests>=2.32.3",
    "sendgrid>=6.12.4",
    "tensorflow>=2.19.0",
]
